Make adjustments to rename the project to bw_projects
closes #4
